requirements.txt
langchain-openai>=0.3.0
openai>=1.0.0
langchain-community>=0.1.2
httpx>=0.24.0